+++ lib/lobbyJ.dart
@@ -11,7 +11,7 @@ import 'package:test8/lobby.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:test8/lobbyO.dart';
 import 'package:test8/lobbyJ.dart';
-import 'package:test8/GameScreenW.dart';
+import 'package:test8/GameScreenA0.dart';
 import 'package:audioplayers/audio_cache.dart';
 import 'package:audioplayers/audioplayers.dart';
 class lobbyJPage extends StatefulWidget {
@@ -39,20 +39,26 @@ class _lobbyJState extends State<lobbyJPage> {
     //this code is to periodically check if GameOpen has been set to true by the room owner, if it is true
     //then move them to gamescreen
     Timer.periodic(Duration(seconds: 2), (timer) async {
-      //print(DateTime.now());
       var sessionQuery = Firestore.instance
           .collection('gameSessions')
-          .where('roomNumber', isEqualTo: _roomNum)
+          .where('roomNumber', isEqualTo: joinedRoom)
           .limit(1);
       var querySnapshot = await sessionQuery.getDocuments();
       var documents = querySnapshot.documents;
       if (documents.length == 0) { /*room doesn't exist? */ return; }
       var isGameOpen = documents[0].data['GameOpen'];
-
+      var docs = await documents[0].reference.collection("players").getDocuments();
+      var firstUser = docs.documents[0].documentID;
       if(isGameOpen == true){
-        Navigator.push(
-            context, MaterialPageRoute(builder: (context) => WaitTimer()));
-        timer.cancel();
+        if(firstUser == currUser){
+          Navigator.push(
+              context, MaterialPageRoute(builder: (context) => MyGame()));
+          timer.cancel();
+        }else{
+          Navigator.push(
+              context, MaterialPageRoute(builder: (context) => AnswerTimer()));
+          timer.cancel();
+        }
       }
     });
   }
@@ -114,7 +120,7 @@ class _lobbyJState extends State<lobbyJPage> {
               _audioCache.play('button.mp3');
               Firestore.instance
                   .collection('gameSessions')
-                  .document(_roomNum)
+                  .document(joinedRoom)
                   .delete();
               Firestore.instance
                   .collection('users')
@@ -210,98 +216,6 @@ class _lobbyJState extends State<lobbyJPage> {
     //x.then((var y)=>z = y.documents.length);
     return z;
     //print(z);
-  }
-
-  void startRoom() {
-    var randNum = new Random();
-    //print(currUser);
-    _roomNum = randNum.nextInt(10000).toString();
-
-    //print(_roomNum);
-    if (currUser != null) {
-      Firestore.instance.collection('gameSessions').document(_roomNum).setData({
-        'roomNumber': _roomNum,
-      });
-      Firestore.instance
-          .collection('gameSessions')
-          .document(_roomNum)
-          .collection('players')
-          .document(currUser)
-          .setData({
-        'question': '',
-        'answer': '',
-      });
-      Firestore.instance
-          .collection('users')
-          .document(currUser)
-          .updateData({'room': _roomNum, 'owner': true});
-    }
-  }
-
-  void joinRoom() async {
-    _roomNum = myController.text;
-
-    //print(getPlayers().then((onValue)=>print(onValue)));
-    // print(getPlayers());
-    //print("ii");
-    //print(_roomNum);
-
-//      List<DocumentSnapshot> templist;
-//
-//      var players = Firestore.instance.collection('gameSessions').getDocuments().toString();
-
-    //List<String> players = (List<String>) Firestore.instance.collection('gameSessions').
-    //List<String> group = (List<String>) document.get("dungeon_group");
-    //var testing =  players.docs.map(doc => doc.data());
-
-//      var list = templist.map((DocumentSnapshot players){
-//        return players.data;
-//      }).toList();
-
-//    List<DocumentSnapshot> templist;
-//    List<Map<dynamic, dynamic>> list = new List();
-//    CollectionReference collectionRef = Firestore.instance.collection(
-//        "gameSessions");
-//    QuerySnapshot collectionSnapshot = await collectionRef.getDocuments();
-//
-//    templist = collectionSnapshot.documents;
-//
-//    list = templist.map((DocumentSnapshot docSnapshot) {
-//      return docSnapshot.data;
-//    }).toList();
-//
-//    var room;
-//
-//    for (var i = 0; i < list.length; i++) {
-//      print(_roomNum);
-//      if (list[i]["roomNumber"] == _roomNum) {
-//        room = list[i];
-//      }
-//    }
-//
-//    int numPlayers = 1;
-//    for (var key in room.keys) {
-//      if (key.startsWith("player")) numPlayers++;
-//    }
-///////////////////////////////////////////
-    //   var playerNumString = "player" + numPlayers.toString();
-
-    if (currUser != null) {
-      Firestore.instance
-          .collection('gameSessions')
-          .document(_roomNum)
-          .collection('players')
-          .document(currUser)
-          .setData({
-        'question': '',
-        'answer': '',
-      });
-      Firestore.instance
-          .collection('users')
-          .document(currUser)
-          .updateData({'room': _roomNum, 'owner': false});
-      print(_roomNum + "ii");
-    }
   }
 
   // this function grabs and returns a list of players in a specified gameSession
